pyproject.toml
[project]
name = "daily-ai-digest"
version = "0.1.0"
description = "U-Chance AI Visual Search POC - AI-powered furniture search using Amazon Bedrock"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "boto3>=1.34.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "streamlit>=1.29.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
]
